*: move "store" to "internal/store"
