+++ etcdserver/api/v2v3/watcher.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/coreos/etcd/clientv3"
 	etcdErr "github.com/coreos/etcd/error"
-	"github.com/coreos/etcd/store"
+	"github.com/coreos/etcd/internal/store"
 )
 
 func (s *v2v3Store) Watch(prefix string, recursive, stream bool, sinceIndex uint64) (store.Watcher, error) {
