+++ etcdserver/apply_v2.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/coreos/etcd/etcdserver/api"
 	"github.com/coreos/etcd/etcdserver/membership"
+	"github.com/coreos/etcd/internal/store"
 	"github.com/coreos/etcd/pkg/pbutil"
-	"github.com/coreos/etcd/store"
 	"github.com/coreos/go-semver/semver"
 )
