+++ etcdserver/membership/cluster.go
@@ -27,12 +27,12 @@ import (
 	"sync"
 	"time"
 
+	"github.com/coreos/etcd/internal/store"
 	"github.com/coreos/etcd/mvcc/backend"
 	"github.com/coreos/etcd/pkg/netutil"
 	"github.com/coreos/etcd/pkg/types"
 	"github.com/coreos/etcd/raft"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/store"
 	"github.com/coreos/etcd/version"
 
 	"github.com/coreos/go-semver/semver"
