+++ etcdserver/membership/store.go
@@ -19,9 +19,9 @@ import (
 	"fmt"
 	"path"
 
+	"github.com/coreos/etcd/internal/store"
 	"github.com/coreos/etcd/mvcc/backend"
 	"github.com/coreos/etcd/pkg/types"
-	"github.com/coreos/etcd/store"
 
 	"github.com/coreos/go-semver/semver"
 )
