+++ internal/store/event.go
