+++ internal/store/event_history.go
