+++ internal/store/node_extern_test.go
