+++ internal/store/stats.go
