+++ internal/store/stats_test.go
