+++ internal/store/store_v2v3_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/coreos/etcd/clientv3"
 	"github.com/coreos/etcd/etcdserver/api/v2v3"
 	"github.com/coreos/etcd/integration"
-	"github.com/coreos/etcd/store"
+	"github.com/coreos/etcd/internal/store"
 
 	"github.com/coreos/pkg/capnslog"
 	"google.golang.org/grpc/grpclog"
