+++ pkg/mock/mockstore/store_recorder.go
@@ -17,8 +17,8 @@ package mockstore
 import (
 	"time"
 
+	"github.com/coreos/etcd/internal/store"
 	"github.com/coreos/etcd/pkg/testutil"
-	"github.com/coreos/etcd/store"
 )
 
 // StoreRecorder provides a Store interface with a testutil.Recorder
